+++ R/signature.R
@@ -1,10 +1,10 @@
 #' Mutational signatures data from COSMIC, the Catalogue Of Somatic Mutations In
-#' Cancer, (v3.1 - June 2020)
+#' Cancer, (v3.2 - March 2021)
 #'
 #' @source Files downloaded from
-#'   \url{https://cancer.sanger.ac.uk/cosmic/signatures/index.tt}, 2021 Feb and
-#'   saved in \code{data-raw/COSMIC.v3.1/data/}. \cr Populated by
-#'   \code{data-raw/COSMIC.v3.1/code/generate-COSMIC.v3.1-genome-sigs.R}.
+#'   \url{https://cancer.sanger.ac.uk/signatures/downloads/}, 2021 Sep and
+#'   saved in \code{data-raw/COSMIC.v3.2/data/}. \cr Populated by
+#'   \code{data-raw/COSMIC.v3.2/code/generate-COSMIC.v3.2-genome-sigs.R}.
 #'
 #' @format A list with a single element, \code{genome}, which is a
 #'   list containing: