+++ R/map_aliquot_ID_to_SP_ID.R
@@ -11,11 +11,11 @@
 #' aliquot.ids <- c("e0fccaf5-925a-41f9-b87c-cd5ee4aecb59", "foo")
 #' SP.ids <- map_aliquot_ID_to_SP_ID(aliquot.ids)
 map_aliquot_ID_to_SP_ID <- function(aliquot.ids) {
-  aliquot.ids.full <- PCAWG.sample.sheet$aliquot_id
+  aliquot.ids.full <- PCAWG7::PCAWG.sample.sheet$aliquot_id
   SP.ids <- sapply(aliquot.ids, function(aliquot.id) {
     if (any(grepl(aliquot.id, aliquot.ids.full))) {
       indices <- grep(aliquot.id, aliquot.ids.full)
-      return(unique(PCAWG.sample.sheet[indices, "icgc_specimen_id"]))
+      return(unique(PCAWG7::PCAWG.sample.sheet[indices, "icgc_specimen_id"]))
     } else {
       return("")
     }})